// frontend/src/api/types.gen.ts
export type JsonDictInput = {
    [key: string]: JsonValueInput;
};

export type JsonDictOutput = {
    [key: string]: JsonValueOutput;
};

export type JsonValueInput = string | number | number | boolean | Array<JsonValueInput> | JsonDictInput | null;

export type JsonValueOutput = string | number | number | boolean | Array<JsonValueOutput> | JsonDictOutput | null;
// ...
export type TagResponse = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Slug
     */
    slug: string;
};

/**
 * TrackedItemCreateRequest
 */
export type TrackedItemCreateRequest = {
    /**
     * Name
     */
    name: string;
    /**
     * Location
     */
    location?: string | null;
    /**
     * Notes
     */
    notes?: string | null;
    attributes: JsonDictInput;
    /**
     * Slug
     */
    slug?: string | null;
};

/**
 * TrackedItemResponse
 */
export type TrackedItemResponse = {
    /**
     * Createdat
     */
    createdAt: string;
    /**
     * Updatedat
     */
    updatedAt: string;
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Slug
     */
    slug: string;
    /**
     * Location
     */
    location?: string | null;
    /**
     * Notes
     */
    notes?: string | null;
    attributes: JsonDictOutput;
};